+++ sdk/core/Azure.Core/src/GeoJson/GeoBoundingBox.cs
@@ -61,8 +61,12 @@ namespace Azure.Core.GeoJson
         }
 
         /// <inheritdoc />
-        public bool Equals(GeoBoundingBox other)
+        public bool Equals(GeoBoundingBox? other)
         {
+            if (other == null)
+            {
+                return false;
+            }
             return West.Equals(other.West) &&
                    South.Equals(other.South) &&
                    East.Equals(other.East) &&
@@ -84,9 +88,9 @@ namespace Azure.Core.GeoJson
         }
 
         /// <summary>
-        ///
+        /// Gets the component of the <see cref="GeoBoundingBox"/> based on it's index.
         /// </summary>
-        /// <param name="index"></param>
+        /// <param name="index">The index of the bounding box component.</param>
         public double this[int index]
         {
             get
